+++ nix/haskell-parsers/parser_tests.nix
@@ -32,6 +32,31 @@ let
         expected = "foo";
       };
     };
+  cabalExecutableTests =
+    let
+      eval = s:
+        let res = parser.parseCabalExecutableNames s; in
+        if res.type == "success" then res.value else res;
+    in
+    {
+      testSimple = {
+        expr = eval ''
+          cabal-version: 3.0
+          name: test-package
+          version: 0.1
+
+          executable foo-exec
+            main-is: foo.hs
+
+          library test
+            exposed-modules: Test.Types
+
+          executable bar-exec
+            main-is: bar.hs
+        '';
+        expected = [ "foo-exec" "bar-exec" ];
+      };
+    };
   # Like lib.runTests, but actually fails if any test fails.
   runTestsFailing = tests:
     let
@@ -44,4 +69,5 @@ in
 {
   "cabal.project" = runTestsFailing cabalProjectTests;
   "package.yaml" = runTestsFailing packageYamlTests;
+  "foo-bar.cabal" = runTestsFailing cabalExecutableTests;
 }