+++ test/simple/flake.nix
@@ -57,8 +57,11 @@
             '';
           };
         };
-        # haskell-flake doesn't set the default package, but you can do it here.
         packages.default = self'.packages.haskell-flake-test;
+
+        # An explicit app to test `nix run .#test` (*without* falling back to
+        # using self.packages.test)
+        apps.app1 = self'.apps.haskell-flake-test;
       };
     };
 }