+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - #134: Add `autoWire` option to control generation of flake outputs
     - #138: Add `checks` to `outputs` submodule
     - #143: Changed `autoWire` to be an enum type, for granular controlling of which outputs to autowire.
+- #137: Expose cabal executables as flake apps. Add a corresponding `outputs.apps` option, while the `outputs.localPackages` option is renamed to `outputs.packages` (it now contains package metadata, including packages and its executables).
 
 ## 0.2.0 (Mar 13, 2023)